fix: Cancel response remaining is u64 instead of scaled decimal value

+++ rust/src/domain/order/client.rs
@@ -180,7 +180,7 @@ pub struct SubmitOrderResponse {
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct CancelSuccess {
     pub order_hash: String,
-    pub remaining: u64,
+    pub remaining: Decimal,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
